created initial /pages structure, moved components loading to pages
