+++ src/pages/ChartPage.tsx
@@ -1,0 +1,6 @@
+// Import: components and models
+import ChartView from "../components/Charts";
+
+export default function ChartPage() {
+  return <ChartView />;
+}
