+++ src/pages/GridPage.tsx
@@ -1,0 +1,6 @@
+// Import: components and models
+import GridFull from "../components/Grids";
+
+export default function GridPage() {
+  return <GridFull />;
+}
