+++ src/pages/TransactionsPage.tsx
@@ -1,0 +1,6 @@
+// Import: components and models
+import TransactionsList from "../components/Transactions";
+
+export default function TransactionsPage() {
+  return <TransactionsList />;
+}
